+++ packages/fluentui/react-northstar/src/components/Popup/Popup.tsx
@@ -201,6 +201,7 @@ export const Popup: React.FC<PopupProps> &
       trapFocus,
       tabbableTrigger,
       trigger: trigger as any,
+      inline,
     }),
     rtl: context.rtl,
   });